requirement.txt
openai
requests
python-dotenv
gradio

